// model/accountTransactionResponse.ts
/**
 * ManageSpace Billing API
 * ManageSpace Billing API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AccountTransactionType } from './accountTransactionType';

export class AccountTransactionResponse {
    /**
    * Account Transaction ID.
    */
    'accountTransactionId': string;
    /**
    * Organization ID.
    */
    'orgId': string;
    /**
    * Site ID.
    */
    'siteId': string;
    /**
    * Account ID.
    */
    'accountId': string;
    /**
    * Account Name
    */
    'accountName': string;
    /**
    * Email
    */
    'email': string | null;
    /**
    * Mobile Number
    */
    'mobileNumber': string;
    /**
    * Transaction Date
    */
    'transactionDate': Date;
    /**
    * Amount Due
    */
    'due': number;
    /**
    * Amount Paid
    */
    'paid': number;
    /**
    * Notes
    */
    'notes': string | null;
    'accountTransactionType': AccountTransactionType;
    /**
    * Invoice Line Item IDs
    */
    'invoiceLineItemIds': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accountTransactionId",
            "baseName": "accountTransactionId",
            "type": "string"
        },
        {
            "name": "orgId",
            "baseName": "orgId",
            "type": "string"
        },
        {
            "name": "siteId",
            "baseName": "siteId",
            "type": "string"
        },
        {
            "name": "accountId",
            "baseName": "accountId",
            "type": "string"
        },
        {
            "name": "accountName",
            "baseName": "accountName",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "mobileNumber",
            "baseName": "mobileNumber",
            "type": "string"
        },
        {
            "name": "transactionDate",
            "baseName": "transactionDate",
            "type": "Date"
        },
        {
            "name": "due",
            "baseName": "due",
            "type": "number"
        },
        {
            "name": "paid",
            "baseName": "paid",
            "type": "number"
        },
        {
            "name": "notes",
            "baseName": "notes",
            "type": "string"
        },
        {
            "name": "accountTransactionType",
            "baseName": "accountTransactionType",
            "type": "AccountTransactionType"
        },
        {
            "name": "invoiceLineItemIds",
            "baseName": "invoiceLineItemIds",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AccountTransactionResponse.attributeTypeMap;
    }
}

export namespace AccountTransactionResponse {
}