// model/createInvoiceLineItemRequest.ts
/**
 * ManageSpace Billing API
 * ManageSpace Billing API Documentation
 *
 * The version of the OpenAPI document: 0.0.11-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CreateInvoiceLineItemRequest {
    /**
    * Account ID
    */
    'accountId': string;
    /**
    * The quantity of items.
    */
    'quantity'?: number;
    /**
    * The price of the unit.
    */
    'unitPrice': number;
    /**
    * A description of the fee.
    */
    'description'?: string;
    /**
    * Product ID
    */
    'productId'?: string;
    /**
    * Rental ID
    */
    'rentalId'?: string;
    /**
    * Rental ID
    */
    'feeId'?: string;
    /**
    * Ledger ID
    */
    'ledgerId'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accountId",
            "baseName": "accountId",
            "type": "string"
        },
        {
            "name": "quantity",
            "baseName": "quantity",
            "type": "number"
        },
        {
            "name": "unitPrice",
            "baseName": "unitPrice",
            "type": "number"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "productId",
            "baseName": "productId",
            "type": "string"
        },
        {
            "name": "rentalId",
            "baseName": "rentalId",
            "type": "string"
        },
        {
            "name": "feeId",
            "baseName": "feeId",
            "type": "string"
        },
        {
            "name": "ledgerId",
            "baseName": "ledgerId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateInvoiceLineItemRequest.attributeTypeMap;
    }
}